+++ colorama/ansi.py
@@ -1,8 +1,8 @@
 # Copyright Jonathan Hartley 2013. BSD 3-Clause license, see LICENSE file.
-'''
+"""
 This module generates ANSI character codes to printing colors to terminals.
 See: http://en.wikipedia.org/wiki/ANSI_escape_code
-'''
+"""
 
 CSI = '\033['
 OSC = '\033]'
@@ -12,11 +12,14 @@ BEL = '\a'
 def code_to_chars(code):
     return CSI + str(code) + 'm'
 
+
 def set_title(title):
     return OSC + '2;' + title + BEL
 
+
 def clear_screen(mode=2):
     return CSI + str(mode) + 'J'
+
 
 def clear_line(mode=2):
     return CSI + str(mode) + 'K'
@@ -34,14 +37,19 @@ class AnsiCodes(object):
 
 
 class AnsiCursor(object):
+
     def UP(self, n=1):
         return CSI + str(n) + 'A'
+
     def DOWN(self, n=1):
         return CSI + str(n) + 'B'
+
     def FORWARD(self, n=1):
         return CSI + str(n) + 'C'
+
     def BACK(self, n=1):
         return CSI + str(n) + 'D'
+
     def POS(self, x=1, y=1):
         return CSI + str(y) + ';' + str(x) + 'H'
 
@@ -93,8 +101,15 @@ class AnsiBack(AnsiCodes):
 class AnsiStyle(AnsiCodes):
     BRIGHT    = 1
     DIM       = 2
+    BRIGHT_OFF  = 21  # Produces double-underline on some terminals.
     NORMAL    = 22
     RESET_ALL = 0
+
+    UNDERLINE     = 4
+    UNDERLINE_OFF = 24
+    REVERSE       = 7
+    REVERSE_OFF   = 27
+
 
 Fore   = AnsiFore()
 Back   = AnsiBack()